feat: use mixin for BGMBox
